requirements.txt
aiohttp==3.7.3
alembic==1.5.0
amqp==2.6.1
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.12.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.9.0
bcrypt==3.2.0
billiard==3.6.3.0
cached-property==1.5.2
cachetools==4.2.0
cattrs==0.9.2
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
cryptography==3.3.1
curlify==2.2.1
defusedxml==0.6.0
dill==0.3.3
dnspython==2.1.0
docutils==0.16
email-validator==1.1.2
fastparquet==0.4.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
flower==0.9.7
funcsigs==1.0.2
future==0.18.2
graphviz==0.16
greenlet==1.0.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.1
grpcio-gcp==0.2.2
gunicorn==19.10.0
hmsclient==0.1.1
httplib2==0.18.1
humanize==3.2.0
idna==2.10
importlib-metadata==3.4.0
iso8601==0.1.13
itsdangerous==1.1.0
JayDeBeApi==1.1.0
Jinja2==2.10.3
joblib==1.0.0
JPype1==1.4.1
json-merge-patch==0.2
jsonschema==3.2.0
kombu==4.6.11
lazy-object-proxy==1.5.2
libcst==0.3.16
llvmlite==0.35.0
lockfile==0.12.2
Mako==1.1.4
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
mock==4.0.3
multidict==5.1.0
mypy-extensions==0.4.3
mysqlclient==1.3.14
natsort==7.1.0
ndg-httpsclient==0.5.1
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==0.25.3
pandas-gbq==0.14.1
paramiko==2.7.2
pendulum==1.4.4
prison==0.1.3
prometheus-client==0.8.0
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.8.0
psycogreen==1.0.2
psycopg2-binary==2.8.6
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pydata-google-auth==1.1.0
Pygments==2.7.4
PyJWT==1.7.1
PyNaCl==1.4.0
pyodbc==4.0.30
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysftp==0.2.9
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.2.0
pytz==2020.5
pytzdata==2020.1
PyYAML==5.3.1
redis==3.2.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
setproctitle==1.2.1
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
sshtunnel==0.1.5
StringDist==1.0.9
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
tornado==5.1.1
tqdm==4.56.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
Werkzeug==0.16.0
WTForms==2.3.3
yarl==1.6.3
zipp==3.4.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.interface==5.2.0
openpyxl==3.0.10
PyHive==0.6.5
sasl==0.3.1
thrift-sasl==0.4.3